bgpd: cosmetic changes for "delete_global_type2_routes"
Two changes -

1) Remove check for `delete_global_type2_routes()`
It is unnecessary and wrong. Even the check failed, it should
continue to delete other types of routes.

2) Remove check for `bgp_dest_has_bgp_path_info_data(rddest)`
It is unnecessary("data" must not be empty) and misleading.
`if (rddest && bgp_dest_has_bgp_path_info_data(rddest))`
Use (locked) node with this check, but unlock with `if (rddest)`,
The mismatched condition is misleading, there seems to be a
mistake to extra unlock.
Just make it clear, immediately exit with `(!rddest)`.

Signed-off-by: anlan_cs <vic.lan@pica8.com>